refactor: apply Rust Specialist improvements (Builder Pattern, Config Module, Error Handling)
